backend/pyproject.toml
[tool.poetry]
name = "classsync"
version = "0.1.0"
description = ""
authors = ["Calvin Laurenson <89622328+calvin-laurenson@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.21"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
onnxruntime = "^1.16.0"
numpy = "^1.26.0"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
